finished CRUDs on the buylist

+++ public/assets/js/admin/buylist.js
@@ -1,11 +1,11 @@
 var currDept;
-window.onload = function() {
+window.onload = function () {
     axios
         .get("/api/getbuylist")
-        .then(function(response) {
+        .then(function (response) {
             buylist = response.data
             console.log(buylist);
-            for(var i = 0; i < buylist.deptList.length; i++) {
+            for (var i = 0; i < buylist.deptList.length; i++) {
                 //create tabs
                 var newTab = $("<li role='presentation'></li>");
                 $(newTab).attr("data-deptId", buylist.deptList[i].id);
@@ -19,9 +19,9 @@ window.onload = function() {
             }
             for (var i = 0; i < buylist.itemList.length; i++) {
                 //create items and sort into respective tables
-                var newTableRow = $("<tr>");
-                $(newTableRow).append("<td>" + buylist.itemList[i].itemName + "</td>");
-                $(newTableRow).append("<td>" + buylist.itemList[i].itemPrice + "</td>");
+                var newTableRow = $("<tr data-itemid='" + buylist.itemList[i].id + "'>");
+                $(newTableRow).append("<td><a href='#' class='delItem'>" + buylist.itemList[i].itemName + "</a></td>");
+                $(newTableRow).append("<td><a href='#' class='updPrice'>" + buylist.itemList[i].itemPrice + "</a></td>");
                 $("#deptId" + buylist.itemList[i].DeptId).append(newTableRow);
             }
 
@@ -31,53 +31,85 @@ window.onload = function() {
             $($("#deptTabs").children()[0]).addClass("active");
             $($("#itemLists").children()[0]).css("display", "table");
 
-            //need to put listener here due to asynchronicity
-            $("#deptTabs > li").on("click", function() {
+            //need to put listeners here due to asynchronicity
+            $("#deptTabs > li").on("click", function () {
                 $("#deptTabs > li").removeClass("active");
                 $(this).addClass("active");
                 currDept = $(this).attr("data-deptId");
                 $("#itemLists > table").css("display", "none");
                 $("#deptTable" + $(this).attr("data-deptId")).css("display", "table");
             });
+            $(".delItem").on("click", function () {
+                if (confirm("Do you want to delete this item?")) {
+                    var itemId = $($($(this).parent()[0]).parent()[0]).attr("data-itemId");
+                    axios.delete("/api/buylist/Item/" + itemId)
+                        .then(function (response) {
+                            console.log(response);
+                            window.location.reload();
+                        })
+                        .catch(function (error) {
+                            console.error(error);
+                        });
+                }
+            });
+            $(".updPrice").on("click", function () {
+                if (confirm("Do you want to update the price of this item?")) {
+                    do {
+                        var newPrice = prompt("Enter a new price: ");
+                    } while (!parseFloat(newPrice));
+                    var itemId = $($($(this).parent()[0]).parent()[0]).attr("data-itemId");
+                    axios.put("/api/buylist/Item", {
+                        itemPrice: newPrice,
+                        id: itemId
+                    }).then(function (response) {
+                        console.log(response);
+                        window.location.reload();
+                    }).catch(function (error) {
+                        console.error(error);
+                    })
+                }
+            });
         })
-        .catch(function(error) {
+        .catch(function (error) {
             console.error(error);
         });
 
-    $("#addDept").on("click", function(e) {
+    $("#addDept").on("click", function (e) {
         e.preventDefault();
         axios.post("/api/buylist/Dept", {
             name: $("#newDept").val().trim()
-        }).then(function(response) {
+        }).then(function (response) {
             console.log(response);
             window.location.reload();
-        }).catch(function(error) {
+        }).catch(function (error) {
             console.error(error);
         });
     });
 
-    $("#remDept").on("click", function(e) {
+    $("#remDept").on("click", function (e) {
         e.preventDefault();
-        axios.delete("/api/buylist/Dept/" + currDept)
-             .then(function(response) {
-                 console.log(response);
-                 window.location.reload();
-             })
-             .catch(function(error) {
-                 console.error(error);
-             });
+        if (confirm("Are you sure you want to delete the current department? It can not be recovered once deleted.")) {
+            axios.delete("/api/buylist/Dept/" + currDept)
+                .then(function (response) {
+                    console.log(response);
+                    window.location.reload();
+                })
+                .catch(function (error) {
+                    console.error(error);
+                });
+        }
     });
 
-    $("#addItem").on("click", function(e) {
+    $("#addItem").on("click", function (e) {
         e.preventDefault();
         axios.post("/api/buylist/Item", {
             itemName: $("#newItem").val().trim(),
             itemPrice: $("#newItemPrice").val().trim(),
             DeptId: currDept
-        }).then(function(response) {
+        }).then(function (response) {
             console.log(response);
-            // window.location.reload();
-        }).catch(function(error) {
+            window.location.reload();
+        }).catch(function (error) {
             console.error(error);
         });
     });
